added keytab for security and pt env
Former-commit-id: 0fd6ba07006064981862754b4691fc5f73e6bc44
Former-commit-id: 95a1dce599cd1c1381349cb38ae6fc435d1987ab

+++ kernel/kernel-fsadapter-hdfs/src/main/java/io/mosip/kernel/fsadapter/hdfs/util/ConnectionUtils.java
@@ -62,14 +62,24 @@ public class ConnectionUtils {
 	 */
 	@Value("${mosip.kernel.fsadapter.hdfs.keytab-file:NOTSET}")
 	private String keytabPath;
-	
+
+	/**
+	 * Field to set connection timeout in milliseconds
+	 */
 	@Value("${mosip.kernel.fsadapter.hdfs.connect.timeout:6000}")
 	private String connectTimeout;
-	
+
+	/**
+	 * Field to set retries the connection after timeout
+	 */
 	@Value("${mosip.kernel.fsadapter.hdfs.connect.max.retries.on.timeouts:10}")
 	private String maxRetries;
-	
 
+	/**
+	 * Field for the kerberos config file location
+	 */
+	@Value("${mosip.kernel.fsadapter.hdfs.krb-file:classpath:krb5.conf}")
+	private String kerberosConfigFile;
 
 	/**
 	 * Field for hadoop FileSystem
@@ -131,10 +141,13 @@ public class ConnectionUtils {
 	private Configuration initSecurityConfiguration(Configuration configuration) throws IOException {
 		configuration.set("dfs.data.transfer.protection", "authentication");
 		configuration.set("hadoop.security.authentication", "kerberos");
-		InputStream krbStream = getClass().getClassLoader().getResourceAsStream("krb5.conf");
-		Path krbPath = Paths.get(hadoopLibPath.toString(), "krb5.conf");
-		Files.copy(krbStream, krbPath);
-		System.setProperty("java.security.krb5.conf", krbPath.toString());
+		Resource confFile = resourceLoader.getResource(kerberosConfigFile);
+		if (confFile.exists()) {
+			InputStream krbStream = confFile.getInputStream();
+			Path krbPath = Paths.get(hadoopLibPath.toString(), "krb5.conf");
+			Files.copy(krbStream, krbPath);
+			System.setProperty("java.security.krb5.conf", krbPath.toString());
+		}
 		UserGroupInformation.setConfiguration(configuration);
 		String user = userName + "@" + kdcDomain;
 		loginWithKeyTab(user, keytabPath);
@@ -215,7 +228,7 @@ public class ConnectionUtils {
 			Files.copy(resource.getInputStream(), keyPath, StandardCopyOption.REPLACE_EXISTING);
 		} else {
 			throw new FSAdapterException(HDFSAdapterErrorCode.KEYTAB_FILE_NOT_FOUND_EXCEPTION.getErrorCode(),
-					HDFSAdapterErrorCode.KEYTAB_FILE_NOT_FOUND_EXCEPTION.getErrorMessage()+": "+keytabPath);
+					HDFSAdapterErrorCode.KEYTAB_FILE_NOT_FOUND_EXCEPTION.getErrorMessage() + ": " + keytabPath);
 		}
 		try {
 			UserGroupInformation.loginUserFromKeytab(user, keyPath.toString());
